Implement integrated server. Used with --app=combined switch. Close #32. Implement command line options override. Can override any config option (--option=value). Add save manipulation to launcher. Add ability to connect to a server from server list. Add ability to start game from world selection screen. Can be started in combined and dedicated server mode. Add numeric scale module.

+++ plugins/voxelman/graphics/plugin.d
@@ -120,8 +120,8 @@ public:
 	override void registerResources(IResourceManagerRegistry resmanRegistry)
 	{
 		auto config = resmanRegistry.getResourceManager!ConfigManager;
-		cameraSensivity = config.registerOption!float("camera_sensivity", 0.4);
-		cameraFov = config.registerOption!float("camera_fov", 60.0);
+		cameraSensivity = config.registerOption!double("camera_sensivity", 0.4);
+		cameraFov = config.registerOption!double("camera_fov", 60.0);
 	}
 
 	override void preInit()
